% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmu202.R
\docType{package}
\name{cmu202-package}
\alias{cmu202}
\alias{cmu202-package}
\title{cmu202: Necessary Packages and Data for 36-202 at Carnegie Mellon University}
\description{
This package provides all necessary pacakges and data for 36-202.
}
\author{
\strong{Maintainer}: Frank Kovacs \email{fkovacs@andrew.cmu.edu}

}
\keyword{internal}
