% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries of the World}
\format{A dataframe with 215 rows and 25 columns.
\describe{
  \item{\code{Country}}{Name of country.}
  \item{\code{LandArea}}{Size in 1000 sq. km.}
  \item{\code{Population}}{Population in millions of people.}
  \item{\code{Density}}{Number of people per square kilometer.}
  \item{\code{GDP}}{Gross Domestic Product (per capita in USD).}
  \item{\code{Rural}}{Percentage of population living in rural areas.}
  \item{\code{CO2}}{CO2 emissions (metric tons per capita).}
  \item{\code{PumpPrice}}{Prive for a liter of gasoline, measured in USD.}
  \item{\code{Military}}{Percentage of government expenditures directed toward the military.}
  \item{\code{Health}}{Percentage of government expenditures directed toward healthcare}
	 \item{\code{ArmedForces}}{Number of active duty military personnel (in 1000s)}
	 \item{\code{Internet}}{Percentage of the population with access to the Internet}
	 \item{\code{Cell}}{Cell phone subscriptions (per 100 people)}
	 \item{\code{HIV}}{Percentage of the population with HIV}
	 \item{\code{Hunger}}{Percentage of the population considered undernourished}
	 \item{\code{Diabetes}}{Percent of the population diagnosed with diabetes}
	 \item{\code{BirthRate}}{Births per 1000 people.}
	 \item{\code{DeathRate}}{Deaths per 1000 people.}
	 \item{\code{ElderlyPop}}{Percentage of the population at least 65 years old}
	 \item{\code{LifeExpectancy}}{Average life expectancy (years)}
	 \item{\code{FemaleLabor}}{Percent of females 15-64 in the labor force}
	 \item{\code{Unemployment}}{Percent of labor force unemployed}
	 \item{\code{EnergyUse}}{Energy Usage}
	 \item{\code{Electricity}}{Electric power consumption (kWh per capita)}
	 \item{\code{Developed}}{Categories for kilowatt hours per capita categorized as: \itemize{
	     \item{\code{1}}{Under 2500}
	     \item{\code{2}}{2500 to 5000}
	     \item{\code{3}}{Over 5000}
	 }}
}}
\usage{
data("countries")
}
\description{
Data DEscribing various countries of the world circa 2004.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Homework 1 Data.
}
\references{
World Development Indicators, World Bank.
}
\keyword{datasets}
