% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{courts}
\alias{courts}
\title{Civil rights case statistics}
\format{A dataframe with 150 rows and 3 columns.
\describe{
  \item{\code{TOTDAM}}{Total amount of damages awarded to plaintiff in USD.}
  \item{\code{BODINJ}}{A binary variable indicating whether a bodily injury was part of the claim.}
  \item{\code{Year}}{Year the lawsuit was filed. Categorized as follows: \itemize{
    \item{1}{Pre-1997}
    \item{2}{1997}
    \item{3}{1998}
    \item{4}{1999}
    \item{5}{2000}
    \item{6}{2001}
    }}
}}
\usage{
data("courts")
}
\description{
Data relating to civil right court cases from the mid-1990s to 
    2001.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Project 2 data.
}
\keyword{datasets}
