% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossMeans.R
\name{crossMeans}
\alias{crossMeans}
\title{Means for 2 way ANOVA}
\usage{
crossMeans(dtf, f1 = "f1", f2 = "f2", y = "y", deci = 1,
  margins = FALSE)
}
\arguments{
\item{dtf}{A data frame.}

\item{f1}{Name of the first categorical predictor variable (quoted or unquoted). Defaults to \code{"f1"}.}

\item{f2}{Name of the second categorical predictor variable (quoted or unquoted). Defaults to \code{"f2"}.}

\item{y}{Name of the quantitative outcome (quoted or unquoted). Defaults to \code{"y"}.}

\item{deci}{Number of decimal places in the output. Defaults to \code{1}.}

\item{margins}{Set to TRUE to add marginal means. Defaults to \code{FALSE}.}
}
\value{
A named \link{matrix} with the resulting 2-way table of means.
}
\description{
A function used to calculate a 2 x 2 table of means for use in 
    ANOVA.
}
\note{
Minor style modifications made by Frank Kovacs (\email{fkovacs AT andrew.cmu.edu}).
}
\references{
Code taken from Howard Seltman's (\email{hseltman AT cmu.edu})
    personal website. \url{http://www.stat.cmu.edu/~hseltman/AboutMe.html}.
}
\author{
Howard Seltman, \email{hseltman AT stat.cmu.edu}
}
