% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Ethics}
\alias{Ethics}
\alias{ethics}
\title{Ethics}
\format{A dataframe with 87 rows and 5 columns.
\describe{
  \item{\code{treatment}}{The hour-long treatment given to each participant.
  \itemize{
    \item{\code{A}}{One-hour ethics tutorial.}
    \item{\code{B}}{A different one-hour ethics tutorial.}
    \item{\code{C}}{Control group.}
    }
  }
  \item{\code{age}}{Age in years.}
  \item{\code{pre}}{Score out of 200 on the pre-test questionnaire.}
  \item{\code{post}}{Score out of 200 on a similar ethics questionnaire taken
  after treatment.}
  \item{\code{diff}}{The change in score, calculated as Post-Pre.}
}}
\usage{
data("ethics")
}
\description{
In the wake of corporate scandals, a researcher wanted to
compare two types of business ethics tutorial (Tutorial ‘A’ and Tutorial ‘B’).
Subjects (students in an MBA program) were obtained (after informed consent).
First, they were all collected in a large lecture hall where they filled
out a well-respected and extensively-designed pencil-and-paper
multiple-choice ‘pretest’ questionnaire to gauge ethical behavior in various
hypothetical scenarios (scored on a 200-point scale). Then, the subjects
were randomized to one of three treatments: Tutorial A
(a one-hour ethics tutorial), Tutorial B (a different one-hour ethics tutorial),
or a control group (who were allowed to sit for the hour and browse the internet).
After the treatment, all the subjects were brought together and made to
take a ‘post-test’ questionnaire on ethics (similar to the pre-test).
The outcome of interest is the difference in scores.
}
\note{
Description text from Gordon Weinberg.

Simplified column names.

36-202 Homework 5 Data.

Obtained from the Spring 2016 edition of 36-202.
}
\keyword{datasets}
