% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Facebook_Grey}
\alias{Facebook_Grey}
\alias{facebook_grey}
\title{Gray Brain Matter and Facebook Friends}
\format{A dataframe with 40 rows and 2 columns.
\describe{
  \item{\code{GMdensity}}{Number of Facebook friends.}
  \item{\code{FBfriends}}{Measure of grey matter density in brain regions of interest. Normalized to z scores.}
}}
\usage{
data("facebook_grey")
}
\description{
Facebook Grey Matter
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Homework Data.
}
\keyword{datasets}
