% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights}
\alias{flights}
\title{Random sample of flight times of United Airlines Flight 433.}
\format{A dataframe with 31 rows and 1 column.
\describe{
  \item{\code{Time}}{The duration of the flight measured in minutes.}
}}
\usage{
data("flights")
}
\description{
Flight times for various iterations of United Flight 433.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

This data set was generated by randomly sampling an existing data set.

36-202 Homework Data.
}
\keyword{datasets}
