% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu}
\alias{flu}
\title{Flu shot prevelance by age.}
\format{A dataframe with 50 rows and 2 columns.
\describe{
  \item{\code{Age}}{Age in years of participant.}
  \item{\code{flu_shot}}{Bool. Did the participant get the flu shot?}
  }}
\usage{
data("flu")
}
\description{
Data of prevalance of flu shot participation.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

\code{flu_shot} is not a factor.

36-202 Lab 10 Data.
}
\keyword{datasets}
