% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hospital}
\alias{Hospital}
\alias{hospital}
\alias{satisfaction}
\title{Patient Satisfaction with Hospitals.}
\format{A dataframe with 23 rows and 5 columns.
\describe{
  \item{\code{Satisfaction}}{Measured on a satisfaction
  index with larger values representing higher satisfaction.}
  \item{\code{Age}}{Age in years of each patient.}
  \item{\code{Severity}}{The severity of illness measured by an index.}
  \item{\code{Anxiety}}{The anxiety level of a patient done on a scale.}
  \item{\code{Gender}}{\itemize{
    \item{0}{Male}
    \item{1}{Female}}
    }
}}
\usage{
data("hospital")
}
\description{
Hospital Satisfaction
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Lab 5 Data.
}
\keyword{datasets}
