% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Houses}
\alias{Houses}
\alias{houses}
\title{Houses}
\format{A dataframe with 30 rows and 6 columns.
\describe{
  \item{\code{Price}}{The most recent list price of the home in thousands of
  U.S. dollars.}
  \item{\code{Area}}{The area of the home in hundreds of square feet.}
  \item{\code{Location}}{The general neighborhood of the home: \itemize{
    \item{\code{suburbs}}
    \item{\code{city}}
    \item{\code{exurb}}{Neither a suburb not a city.}
    }
  }
  \item{\code{dum.loc1}}{One of two dummy variables for coding location.}
  \item{\code{dum.loc2}}{The second dummy variable for coding location.}
}}
\usage{
data("houses")
}
\description{
Based on recent data, a variety of variables on regional houses
are available.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Description and most variable text from Gordon Weinberg.

Changed variables names to a single, uppercase word.

36-202 Data.

Obtained from the Spring 2016 edition of 36-202.
}
\keyword{datasets}
