% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mammals}
\alias{Mammals}
\alias{mammals}
\alias{MammalSleep}
\title{Sleep, Gestation and Danger of Mammals}
\format{A dataset containing 51 rows and 6 columns.
\describe{
  \item{\code{Animal}}{Animal Name}
  \item{\code{Sleep}}{Hours of sleep in a day.}
  \item{\code{BodyWt}}{Weight in kg.}
  \item{\code{LifeSpan}}{Maximum life span in years.}
  \item{\code{Gestation}}{The gestation period of the animal.}
  \item{\code{Danger}}{The danger rating of an animal on a scale from 1-5.}
  }}
\usage{
data("mammals")
}
\description{
Mammals
}
\note{
36-202 Lab Data.
}
\keyword{datasets}
