% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nflDraft}
\alias{nflDraft}
\alias{NFL}
\title{2000 NFL Draft Data}
\format{A dataframe with 40 rows and 6 columns.
\describe{
\item{\code{First}}{First name of the prospect.}
\item{\code{Last}}{Last name of the prospect.}
\item{\code{Position}}{\itemize{
  \item{Wide Receiver}
  \item{Guard}
  \item{Offensive Tackle}
  }}
\item{\code{Weight}}{Weight in pounds.}
\item{\code{40Dash}}{Time to run 40 yards measured in seconds.}
\item{\code{Rating}}{Rating as a prospect on a scale of 10.}
}}
\usage{
data("nflDraft")
}
\description{
A dataset containing information about 40 draft prospects from the 2000 NFL Draft.
}
\note{
36-202 Lab Data.
}
\author{
Gordon Weinberg \email{gordonw@stat.cmu.edu}
}
\keyword{datasets}
