% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrition}
\alias{nutrition}
\title{Nutrition Data}
\format{A dataframe with 340 rows and 4 columns.
\describe{
  \item{\code{Shrt_Desc}}{The name of the food.}
  \item{\code{Energy}}{Food energy in kcal per 100 grams.}
  \item{\code{Water}}{Water in grams per 100 grams of food. Itemized as follows: \itemize{
    \item{\code{1}}{Between 0 and <25 grams}
    \item{\code{2}}{Between 25 and <50 grams}
    \item{\code{3}}{Between 50 and <75 grams}
    \item{\code{4}}{Between 75 and <100 grams}
    }}
  \item{\code{Choles}}{Cholesterol values. Classified as \code{0} if 0mg / 100g, \code{1} otherwise.}
  \item{\code{Shrt_Desc}}{A short description of the item.}
}}
\usage{
data("nutrition")
}
\description{
Data from the United States Department of Agriculture's Natoinal
    Nutrient Database for Standard Reference. The data given was collected during a one-week interval in 2017.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Project 2 data.
}
\keyword{datasets}
