% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc}
\alias{nyc}
\title{New York City Housing}
\format{A dataframe with 339 rows and 3 columns.
\describe{
  \item{\code{Income}}{Total household income in USD.}
  \item{\code{Gender}}{Respondent's self-reported gender identity: \itemize{
    \item{\code{1}}{Female}
    \item{\code{2}}{Male}
    }}
  \item{\code{Gender}}{Respondent's categorical self-assesment of health status. Higher values indicate a more healthy individual.}
}}
\usage{
data("nyc")
}
\description{
Data from  the New York City Housing and Vacancy Survey to understand housing conditions.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Project 2 data.
}
\keyword{datasets}
