% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Offer}
\alias{Offer}
\alias{offer}
\title{Offer}
\format{A dataframe with 36 rows and 3 columns.
\describe{
  \item{\code{Offer}}{Amount offered in hundreds of dollars.}
  \item{\code{Age}}{The age category of the seller: \itemize{
    \item{\code{1}}{Young}
    \item{\code{2}}{Middle-Aged}
    \item{\code{3}}{Elderly}
    }}
  \item{\code{Gender}}{\itemize{
    \item{\code{M}}{Male}
    \item{\code{F}}{Female}
    }}
}}
\usage{
data("offer")
}
\description{
Data from a consumer advocacy organization concerning the amount of
money offered by dealerships for used cars as a function of the seller's age and
gender.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Lab Data.
}
\keyword{datasets}
