% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqn.R
\name{qqn}
\alias{qqn}
\title{Quantile-Normal Plot with Limits}
\usage{
qqn(x, ylab = deparse(substitute(x)), lty1 = 1, lty2 = 1, swap = FALSE,
  conf = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector of data for which to create the QQ plot.}

\item{ylab}{A \code{string} indicating the label for the y-axis of the 
plot. Defaults to the variable name of the argument \code{x}.}

\item{lty1}{The line type for the reference line. See \link{par} for 
more information. Defaults to \code{1}.}

\item{lty2}{The line types for the confidence intervals. See \link{par}
for more information. Defaults to \code{1}.}

\item{swap}{A boolean value indicating whether to swap the x and y axes. 
If set to \code{TRUE}, the quantiles of standard normal will be placed on 
the Y axis. Defaults to \code{FALSE}.}

\item{conf}{A boolean value indicating whether to produce the confidence 
bands. Defaults to \code{TRUE}.}

\item{...}{Other arguments passed to \link{qqnorm}.}
}
\value{
An invisible list with the following components: \describe{
  \item{\code{z}}{The normal quantiles.}
  \item{\code{x}}{The input data.}
  \item{\code{slope}}{The slope of the plot.}
  \item{\code{zz}}{The z values sorted in ascending order.}
  \item{\code{P}}{Probability points. See \link{ppoints} for more 
      information.}
  \item{\code{SE}}{The delta-method standard error for the confidence 
      interval.}
}.
Additionally the function will always plot the normal QQ plot.
}
\description{
A function used to calculate a normal QQ plot with confidence 
    bands representing a 95% confidence interval for the acceptable range
    of values. This function extends the graphing capabilities of base R and 
    implicitly creates a plot that may be added to.
}
\note{
All documentation and code taken from Brian Junker and Howard Seltman.
}
\references{
Code taken from Howard Seltman's (\email{hseltman AT cmu.edu})
    personal website. \url{http://www.stat.cmu.edu/~hseltman/AboutMe.html}. 
    Code may be found at \url{http://www.stat.cmu.edu/~hseltman/files/qqn.R}.

To read more about the delta method of bound calculations, please
    see \url{https://genome.sph.umich.edu/wiki/Code_Sample:_Generating_QQ_Plots_in_R}
}
\author{
Brian Junker, \email{brian AT stat.cmu.edu}
}
