% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R202.R
\docType{package}
\name{R202-package}
\alias{R202}
\alias{R202-package}
\title{R202: Necessary Packages and Data for 36-202 at Carnegie Mellon University}
\description{
This package provides all necessary pacakges and data for 36-202.
}
\author{
\strong{Maintainer}: Frank Kovacs \email{fkovacs@andrew.cmu.edu}

}
\keyword{internal}
