% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{social}
\alias{social}
\title{Social Media}
\format{A dataframe with 500 rows and 3 columns.
\describe{
  \item{\code{shares}}{The number of shares in a social network.}
  \item{\code{videos}}{The number of videos in the article. Categories: \itemize{
    \item{\code{0}}{0}
    \item{\code{1}}{1}
    \item{\code{2}}{>=2}
    }}
  \item{\code{channel}}{The type of channel. One of: \itemize{
    \item{\code{Business}}{}
    \item{\code{Entertainment}}{}
    \item{\code{Lifestyle}}{}
    \item{\code{Other}}{}
    \item{\code{SocialMedia}}{}
    \item{\code{Tech}}{}
    \item{\code{World}}{}
    }}
}}
\usage{
data("social")
}
\description{
Data from Predicting the Popularity of Online News.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Project 2 data.
}
\keyword{datasets}
