% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swim}
\alias{swim}
\title{Swimmer Confidence}
\format{A dataframe with 16 rows and 2 columns.
\describe{
  \item{\code{Pessimism}}{The pessimism score of each swimmer. A high score
  indicates a more pessimistic mindset.}
  \item{\code{Ratio}}{The ratio of swimming times for each trial -- trial1/trial2.
  An \code{Ratio} value > 1 indicates that a swimmer performed better on the second
  trial and vice-versa.}
}}
\source{
Author Seligman, M.E.P., Nolen-Hoeksema, S., Thornton, N. and Thornton, C.M. (1990)
Explanatory style as a mechanism of disappointing athletic performance.
\emph{Psychological Science}, \bold{1}, 143--146,
\url{http://www.uvm.edu/~dhowell/StatPages/Swimming/Swimming.html}
}
\usage{
data("swim")
}
\description{
You often hear that a negative outlook can be self-defeating. To
test whether this is really true, psychology researchers conducted a study in
1990 on a group of 33 swim-team members at the University of California at
Berkeley. 16 of them are cataloged in this data set.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-201 Homework 2 Data.

36-202 Lab 2 Data.

Obtained from the Fall 2015 edition of 36-201.
}
\keyword{datasets}
