% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_list.R
\name{vl}
\alias{vl}
\title{Produce a Mapping of Available Vignettes with Titles}
\usage{
vl()
}
\value{
\code{Named List}. A named list containing attributes about the
    available vignettes. For simplicity, it is recommended that one only
    calls the `vl()` function in an R terminal.
}
\description{
This function provides a wrapper for students to see the available
    vignettes in the \code{cmu202} package. \code{vl} opens a tab in the 
    primary editor pane of RStudio displaying the vignettes and titles available
    in the package.
}
\note{
This command produces the same result as 
    \code{vignette(package = "cmu202")}
}
\examples{
# vl()

}
\seealso{
\link{vignette}, \link{vv}, \link{lv}
}
