% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_vignette.R
\name{vv}
\alias{vv}
\title{View a Vignette in the cmu202 Package}
\usage{
vv("vignetteName")
}
\arguments{
\item{\code{vignetteName}}{A \code{string} value indicating the name of the 
vignette you wish to see. Will return an \code{error} if the argument does 
not match any existing vignettes.}
}
\value{
\code{vignette}. An object of class \code{vignette}. If one simply wishes
    to view the vignette, calling the function without assignment will 
    will display the vignette in the "viewer" tab.
}
\description{
This function provides a wrapper for students to view vignettes
    only in the \code{cmu202} package. \code{view_vignette} is intended to make
    the viewing of vignettes as painless as humanly possibe. Upon calling 
    the function, the vignette will become viewable in the "viewer" tab of the 
    "files" pane in RStudio.
}
\note{
This command is identical to \code{vignette("vignetteName", package = "cmu202")}
}
\examples{
# view_vignette("egg") ## Views the egg vignette.
}
\seealso{
\link{vignette}, \link{vl}, \link{lv}
}
