% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Airlines}
\alias{Airlines}
\alias{airlines}
\title{Airlines}
\format{A dataframe with 500 rows and 3 columns.
\describe{
 \item{\code{DayofWeek}}{Day of the week (1 = Monday; 7 = Sunday)}
 \item{\code{ArrDelay}}{Arrival delay in minutes (negative means an early arrival)}
 \item{\code{LateAircraft}}{Was there a delay due to a late aircraft? 1 = Yes 0 = No}
}}
\usage{
data("airlines")
}
\description{
A dataset containing federally tracked data for flights. See TO-DO: INSERT INFO FUNCTION
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Initially used in 36-200.

Project 2 data.
}
\keyword{datasets}
