% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Bikes}
\alias{Bikes}
\alias{bikes}
\title{Bikes}
\format{A dataframe with 951 rows and 3 columns.
\describe{
	 \item{\code{Day}}{The day of the week of the day.}
	 \item{\code{Weather}}{The type of weather on that day. Coded with respect to the following values: 1 = clear, few clouds, partly cloudy, 2 = mist & cloudy, mist & broken clouds, mist & few clouds, mist, 3 = light snow, light rain & Thunderstorm & scattered clouds, light rain & scattered clouds, 4 = heavy rain & ice pellets & thunderstorm & mist, snow & fog}
	 \item{\code{Casual}}{The number of casual bike users}
}}
\usage{
data("bikes")
}
\description{
A dataset information about bicycle rentals for days in the years 2011 and 2012.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

Project 2 data.
}
\keyword{datasets}
