% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmu202.R
\docType{package}
\name{cmu202-package}
\alias{cmu202}
\alias{cmu202-package}
\title{cmu202: Necessary Packages and Data for 36-202 at Carnegie Mellon University}
\description{
This package provides all necessary pacakges and data for 36-202.
}
\author{
\strong{Maintainer}: Francis Kovacs \email{fkovacs@andrew.cmu.edu} [owner, programmer]

Other contributors:
\itemize{
  \item Rebecca Nugent \email{rnugent@stat.cmu.edu} [data contributor, expert, thesis advisor]
  \item Gordon Weinberg \email{gordonw@andrew.cmu.edu} [data contributor, teacher]
  \item Kayla Frisoli \email{kfrisoli@andrew.cmu.edu} [writer of added text]
  \item Brian Junker \email{brian@stat.cmu.edu} (Wrote the qqn function with Howard Seltman) [author in quotations or text abstracts]
  \item Howard Seltman \email{howard@stat.cmu.edu} (Wrote the qqn function with Brian Junker. Wrote the crossMeans function.) [author in quotations or text abstracts]
}

}
\keyword{internal}
