% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Depression}
\alias{Depression}
\alias{depression}
\title{Depression after an earthquake.}
\format{A dataframe with 125 rows and 3 columns.
\describe{
  \item{\code{Dep_Score}}{The student's depression score (higher -> more depressed)}
  \item{\code{Week}}{Weeks relative to the earthquake. \itemize{
    \item{-2}{Two weeks prior to the earthquake.}
    \item{1}{One week after the earthquake.}
    \item{4}{Four weeks after the earthquake.}
    \item{7}{Seven weeks after the earthquake.}
    \item{10}{Ten weeks after the earthquake.}
    }}
  \item{\code{Subject}}{Subject Number.}
}}
\usage{
data("depression")
}
\description{
Depression
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Lab Data.
}
\author{
Gordon Weinberg \email{gordonw@stat.cmu.edu}
}
\keyword{datasets}
