% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facebook_grey}
\alias{facebook_grey}
\title{Gray Brain Matter and Facebook Friends}
\format{A dataframe with 40 rows and 2 columns.
\describe{
  \item{\code{GMdensity}}{Number of Facebook friends.}
  \item{\code{FBfriends}}{Measure of grey matter density in brain regions of interest. Normalized to z scores.}
}}
\usage{
data("facebook_grey")
}
\description{
Data collected from a study relating social media interaction
    and brain structure.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Homework Data.
}
\references{
Kanai et al. "Online social network size is reflected in human 
    brain structure." Proceedings of the Royal Society, 2012: 279(1732): 
    1327-1334.
}
\keyword{datasets}
