% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_vignette.R
\name{lv}
\alias{lv}
\title{List Vignettes available in the R202 Package}
\usage{
lv()
}
\arguments{
\item{None}{}
}
\value{
\code{vector}. A vector of \code{string}s indicating the 
    vignettes available in the R202 package.
}
\description{
This function provides a wrapper for students to list the available
    vignettes only in the \code{R202} package. \code{lv} returns a \code{vector} 
    of vignettes a sstrings, each of which may be passed to the \code{vv()} 
    function.
}
\note{
This command manipulates the datafame returned by 
    \code{vignette(package = "R202")}
}
\examples{
# x <- lv()
## x is a vector containing the vignettes
## the below command will be valid
# vv(lv()[[1]])
}
\seealso{
\link{vignette}, \link{vv}, \link{vl}
}
