% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp}
\alias{temp}
\title{Power Load by High Temperature}
\format{A dataframe with 25 rows and 2 columns.
\describe{
  \item{\code{temp}}{The daily high temperature recorded for the company.}
  \item{\code{load}}{The corresponding daily peak power load.}
  }}
\usage{
data("temp")
}
\description{
Data relating to the load of a company's power input and 
    temperature.
}
\note{
The documentation does not auto-update. If one changes the base file,
  this documentation will no longer remain accurate.

36-202 Lab 3 Data.
}
\keyword{datasets}
